+++ src/oci/database/models/create_autonomous_database_base.py
@@ -318,7 +318,10 @@ class CreateAutonomousDatabaseBase(object):
     def db_workload(self):
         """
         Gets the db_workload of this CreateAutonomousDatabaseBase.
-        The Autonomous Database workload type. OLTP indicates an Autonomous Transaction Processing database and DW indicates an Autonomous Data Warehouse. The default is OLTP.
+        The Autonomous Database workload type. The following values are valid:
+
+        - OLTP - indicates an Autonomous Transaction Processing database
+        - DW - indicates an Autonomous Data Warehouse database
 
         Allowed values for this property are: "OLTP", "DW"
 
@@ -332,7 +335,10 @@ class CreateAutonomousDatabaseBase(object):
     def db_workload(self, db_workload):
         """
         Sets the db_workload of this CreateAutonomousDatabaseBase.
-        The Autonomous Database workload type. OLTP indicates an Autonomous Transaction Processing database and DW indicates an Autonomous Data Warehouse. The default is OLTP.
+        The Autonomous Database workload type. The following values are valid:
+
+        - OLTP - indicates an Autonomous Transaction Processing database
+        - DW - indicates an Autonomous Data Warehouse database
 
 
         :param db_workload: The db_workload of this CreateAutonomousDatabaseBase.
@@ -678,7 +684,9 @@ class CreateAutonomousDatabaseBase(object):
     def nsg_ids(self):
         """
         Gets the nsg_ids of this CreateAutonomousDatabaseBase.
-        A list of the `OCIDs`__ of the network security groups (NSGs) that this DB system belongs to. Setting this to an empty array after the list is created removes the resource from all NSGs. For more information about NSGs, see `Security Rules`__.
+        A list of the `OCIDs`__ of the network security groups (NSGs) that this resource belongs to. Setting this to an empty array after the list is created removes the resource from all NSGs. For more information about NSGs, see `Security Rules`__.
+        **NsgIds restrictions:**
+        - Autonomous Databases with private access require at least 1 Network Security Group (NSG). The nsgIds array cannot be empty.
 
         __ https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm
         __ https://docs.cloud.oracle.com/Content/Network/Concepts/securityrules.htm
@@ -693,7 +701,9 @@ class CreateAutonomousDatabaseBase(object):
     def nsg_ids(self, nsg_ids):
         """
         Sets the nsg_ids of this CreateAutonomousDatabaseBase.
-        A list of the `OCIDs`__ of the network security groups (NSGs) that this DB system belongs to. Setting this to an empty array after the list is created removes the resource from all NSGs. For more information about NSGs, see `Security Rules`__.
+        A list of the `OCIDs`__ of the network security groups (NSGs) that this resource belongs to. Setting this to an empty array after the list is created removes the resource from all NSGs. For more information about NSGs, see `Security Rules`__.
+        **NsgIds restrictions:**
+        - Autonomous Databases with private access require at least 1 Network Security Group (NSG). The nsgIds array cannot be empty.
 
         __ https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm
         __ https://docs.cloud.oracle.com/Content/Network/Concepts/securityrules.htm